backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.15
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
pydantic-settings==2.1.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0